moved create plan inside sidebar, got rid of secondary list items, removed money box
