+++ my-app/src/components/Student/Dashboard/StudentAppSideBar.js
@@ -1,6 +1,6 @@
-import React from 'react';
+import React from "react";
 import clsx from 'clsx';
-import { makeStyles } from '@material-ui/core/styles';
+import { makeStyles } from "@material-ui/core/styles";
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Drawer from '@material-ui/core/Drawer';
 import Box from '@material-ui/core/Box';
@@ -18,25 +18,7 @@ import Link from '@material-ui/core/Link';
 import MenuIcon from '@material-ui/icons/Menu';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
 import NotificationsIcon from '@material-ui/icons/Notifications';
-import { mainListItems, secondaryListItems } from './Dashboard/listItems';
-
-import Trainers from "./Dashboard/Trainers";
-import Money from "./Dashboard/Money";
-import StudentTable from "./Dashboard/Table";
-import Chart from "./Dashboard/Chart";
-
-function Copyright() {
-  return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {"Copyright © "}
-      <Link color="inherit" href="https://material-ui.com/">
-        Your Website
-      </Link>{" "}
-      {new Date().getFullYear()}
-      {"."}
-    </Typography>
-  );
-}
+import { mainListItems, secondaryListItems } from './listItems';
 
 const drawerWidth = 240;
 
@@ -119,7 +101,7 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-export default function StudentDashboard(props) {
+export default function TrainerAppSideBar() {
   const classes = useStyles();
   const [open, setOpen] = React.useState(true);
   const handleDrawerOpen = () => {
@@ -132,49 +114,50 @@ export default function StudentDashboard(props) {
 
   const cards = [1, 2, 3, 4, 5, 6];
   return (
-    <div
-     
-      className={classes.root}>
-      <CssBaseline />
-      <AppBar
-        style={{
-          marginTop: "5%"
-        }}
-        position="absolute"
-        className={clsx(classes.appBar, open && classes.appBarShift)}
-      >
-        <Toolbar
-     
-        className={classes.toolbar}>
-          <IconButton
-            edge="start"
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            className={clsx(
-              classes.menuButton,
-              open && classes.menuButtonHidden
-            )}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Typography
-            component="h1"
-            variant="h6"
-            color="inherit"
-            noWrap
-            className={classes.title}
-          >
-            {props.pagename}
-          </Typography>
-          <IconButton color="inherit">
-            <Badge badgeContent={4} color="secondary">
-              <NotificationsIcon />
-            </Badge>
-          </IconButton>
-        </Toolbar>
-      </AppBar>
-      <Drawer
+    <div className={classes.root} >
+      {/* <Students /> */}
+
+      {/* <StudentTable /> */}
+
+
+
+  <AppBar
+  style={{
+    marginTop: "5%"
+  }}
+  position="absolute"
+  className={clsx(classes.appBar, open && classes.appBarShift)}
+  >
+  <Toolbar className={classes.toolbar}>
+    <IconButton
+      edge="start"
+      color="inherit"
+      aria-label="open drawer"
+      onClick={handleDrawerOpen}
+      className={clsx(
+        classes.menuButton,
+        open && classes.menuButtonHidden
+      )}
+    >
+      <MenuIcon />
+    </IconButton>
+    <Typography
+      component="h1"
+      variant="h6"
+      color="inherit"
+      noWrap
+      className={classes.title}
+    >
+      Student(Insert Name Here) Dashboard
+    </Typography>
+    <IconButton color="inherit">
+      <Badge badgeContent={4} color="secondary">
+        <NotificationsIcon />
+      </Badge>
+    </IconButton>
+  </Toolbar>
+  </AppBar>
+  <Drawer
         variant="permanent"
         classes={{
           paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose),
@@ -186,39 +169,12 @@ export default function StudentDashboard(props) {
             <ChevronLeftIcon />
           </IconButton>
         </div>
-        <Divider />
+        {/* <Divider /> */}
         <List>{mainListItems}</List>
-        <Divider />
-        <List>{secondaryListItems}</List>
+        {/* <Divider /> */}
+        {/* <List>{secondaryListItems}</List> */}
       </Drawer>
-      <main className={classes.content}>
-        <div className={classes.appBarSpacer} />
-        <Container maxWidth="lg" className={classes.container}>
-          <Grid container spacing={3}>
-            {/* Chart on weight */}
-            <Grid item xs={12} md={8} lg={9}>
-              <Paper className={fixedHeightPaper}>
-                <Chart />
-              </Paper>
-            </Grid>
-            {/* Available Funds */}
-            {/* <Grid item xs={12} md={4} lg={3}> */}
-              {/* <Paper className={fixedHeightPaper}> */}
-                {/* <Money /> */}
-              {/* </Paper> */}
-            {/* </Grid> */}
-            {/* Recent Stats */}
-            <Grid item xs={12}>
-              <Paper className={classes.paper}>
-                <StudentTable />
-              </Paper>
-            </Grid>
-          </Grid>
-          <Box pt={4}>
-            <Copyright />
-          </Box>
-        </Container>
-      </main>
-    </div>
-  );
+</div>
+
+)
 }
