+++ my-app/src/components/Student/StudentProfile.js
@@ -1,97 +1,144 @@
 import React from "react";
+import clsx from 'clsx';
 import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
 import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
+import Link from '@material-ui/core/Link';
+import Profile from "./Dashboard/Profile";
+import CssBaseline from '@material-ui/core/CssBaseline';
+import Grid from '@material-ui/core/Grid';
+import Container from '@material-ui/core/Container';
+import Paper from '@material-ui/core/Paper';
+import StudentAppSideBar from "./Dashboard/StudentAppSideBar"
+import Box from '@material-ui/core/Box';
 
-// import Avatar from "@material-ui/core/Avatar";
+function Copyright() {
+  return (
+    <Typography variant="body2" color="textSecondary" align="center">
+      {"Copyright © "}
+      <Link color="inherit" href="https://material-ui.com/">
+        Your Website
+      </Link>{" "}
+      {new Date().getFullYear()}
+      {"."}
+    </Typography>
+  );
+}
+
 const currentUser = {
   isStudent: false
 };
 
-const useStyles = makeStyles({
+const useStyles = makeStyles(theme => ({
   root: {
-    width: "20%",
-    marginTop: "4%",
-    border: "none"
+    display: "flex"
   },
-
+  toolbar: {
+    paddingRight: 24 // keep right padding when drawer closed
+  },
+  toolbarIcon: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "flex-end",
+    padding: "0 8px",
+    ...theme.mixins.toolbar
+  },
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    })
+  },
+  appBarShift: {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  menuButton: {
+    marginRight: 36
+  },
+  menuButtonHidden: {
+    display: "none"
+  },
   title: {
-    fontSize: "1.4rem"
+    flexGrow: 1
   },
-  pos: {
-    marginBottom: 12
+  drawerPaper: {
+    position: "relative",
+    whiteSpace: "nowrap",
+    width: drawerWidth,
+    transition: theme.transitions.create("width", {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen
+    })
   },
-  pic: {
-    width: 200,
-    padding: "2%"
+  drawerPaperClose: {
+    overflowX: "hidden",
+    transition: theme.transitions.create("width", {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    }),
+    width: theme.spacing(7),
+    [theme.breakpoints.up("sm")]: {
+      width: theme.spacing(9)
+    }
+  },
+  appBarSpacer: theme.mixins.toolbar,
+  content: {
+    flexGrow: 1,
+    height: "100vh",
+    overflow: "auto"
+  },
+  container: {
+    paddingTop: theme.spacing(4),
+    paddingBottom: theme.spacing(4)
+  },
+  paper: {
+    padding: theme.spacing(2),
+    display: "flex",
+    overflow: "auto",
+    flexDirection: "column"
+  },
+  fixedHeight: {
+    height: 240
   }
-});
+}));
+
+const drawerWidth = 240;
 
 export default function StudentProfile(props) {
   const classes = useStyles();
+  const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
+
 
   return (
-    <Card className={classes.root} variant="outlined">
-      <CardContent
-        style={{
-          marginTop: "5%",
-          marginLeft: "10%",
-          backgroundColor: "rgba(130, 143, 211, 0.25)",
-          padding: "2%"
-        }}
-      >
-        <Typography
-          className={classes.title}
-          color="textSecondary"
-          gutterBottom
-          style={{
-            textAlign: "center",
-            textTransform: "uppercase"
-          }}
-        >
-          Student profile
-        </Typography>
-        <Typography variant="h5" component="h2" color="primary" align="left">
-          {props.studentData.name}
-        </Typography>
-        <img
-          className={classes.pic}
-          src={props.studentData.avatar}
-          alt="Smiley face"
-        />
-        {/* <Avatar aria-label="recipe" className={classes.avatar}></Avatar> */}
+    <div className={classes.root} > 
+    <CssBaseline />
+    <StudentAppSideBar />
+    <main className={classes.content}>
+      <div className={classes.appBarSpacer} />
+      <Container maxWidth="lg" className={classes.container}>
+        <Grid container spacing={10}>
+          {/* Students Images */}
+          <Grid item xs={12} md={12} lg={20}>
+            <Paper className={fixedHeightPaper}>
+              <Profile studentData={props.studentData} />
+            </Paper>
+          </Grid>          
+        </Grid>
+        <Box pt={4}>
+          <Copyright />
+        </Box>
+      </Container>
+    </main>
+  </div>
 
-        <Typography className={classes.pos} variant="h5" color="Secondary">
-          Goal : {props.studentData.goal}
-        </Typography>
-        <Typography className={classes.pos} color="textSecondary">
-          Height : {props.studentData.height}
-        </Typography>
-        <Typography className={classes.pos} color="textSecondary">
-          Weight : {props.studentData.weight}
-        </Typography>
-        <Typography className={classes.pos} color="textSecondary">
-          BMI :{" "}
-          {(
-            (props.studentData.weight /
-              (props.studentData.height *
-                0.39 *
-                props.studentData.height *
-                0.39)) *
-            703
-          ).toFixed(2)}
-        </Typography>
-      </CardContent>
-      {currentUser.isStudent ? (
-        <CardActions>
-          <Button size="small">update profile</Button>
-        </CardActions>
-      ) : (
-        ""
-      )}
-    </Card>
   );
 }
