+++ my-app/src/components/Trainer/TrainerProfile.js
@@ -1,20 +1,19 @@
-import React, { useState } from "react";
+import React from "react";
 import clsx from 'clsx';
 import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Container from '@material-ui/core/Container';
 import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
 import Link from '@material-ui/core/Link';
+import ProfileTr from "./Dashboard/ProfileTr";
+import CssBaseline from '@material-ui/core/CssBaseline';
 import Grid from '@material-ui/core/Grid';
+import Container from '@material-ui/core/Container';
 import Paper from '@material-ui/core/Paper';
 import TrainerAppSideBar from "./Dashboard/TrainerAppSidebar"
 import Box from '@material-ui/core/Box';
-
-// import Avatar from '@material-ui/core/Avatar';
 
 function Copyright() {
   return (
@@ -29,85 +28,117 @@ function Copyright() {
   );
 }
 
-const useStyles = makeStyles({
-  root: {
-    minWidth: 200
-  },
+const currentUser = {
+  isStudent: false
+};
 
+const useStyles = makeStyles(theme => ({
+  root: {
+    display: "flex"
+  },
+  toolbar: {
+    paddingRight: 24 // keep right padding when drawer closed
+  },
+  toolbarIcon: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "flex-end",
+    padding: "0 8px",
+    ...theme.mixins.toolbar
+  },
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    })
+  },
+  appBarShift: {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  menuButton: {
+    marginRight: 36
+  },
+  menuButtonHidden: {
+    display: "none"
+  },
   title: {
-    fontSize: 14
+    flexGrow: 1
   },
-  pos: {
-    marginBottom: 12
+  drawerPaper: {
+    position: "relative",
+    whiteSpace: "nowrap",
+    width: drawerWidth,
+    transition: theme.transitions.create("width", {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen
+    })
   },
-  pic: {
-    width: 400
+  drawerPaperClose: {
+    overflowX: "hidden",
+    transition: theme.transitions.create("width", {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    }),
+    width: theme.spacing(7),
+    [theme.breakpoints.up("sm")]: {
+      width: theme.spacing(9)
+    }
   },
-  desc: {
-    width: 400
+  appBarSpacer: theme.mixins.toolbar,
+  content: {
+    flexGrow: 1,
+    height: "100vh",
+    overflow: "auto"
+  },
+  container: {
+    paddingTop: theme.spacing(4),
+    paddingBottom: theme.spacing(4)
+  },
+  paper: {
+    padding: theme.spacing(2),
+    display: "flex",
+    overflow: "auto",
+    flexDirection: "column"
+  },
+  fixedHeight: {
+    height: 240
   }
-});
+}));
+
+const drawerWidth = 240;
 
 export default function TrainerProfile(props) {
-  console.log(props);
-
   const classes = useStyles();
   const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
 
 
   return (
-    <div className={classes.root} >
-      <CssBaseline /> 
-      <TrainerAppSideBar />
-      <main className={classes.content}>
-        <div className={classes.appBarSpacer} />
-        <Container maxWidth="lg" className={classes.container}>
-          <Grid container spacing={3}>
-            {/* Trainer Profile */}
-            <Grid item xs={10} md={10} lg={10}>
-              <Paper className={fixedHeightPaper}>
-                      <Card className={classes.root} variant="outlined">
-                <CardContent>
-                  <Typography
-                    className={classes.title}
-                    color="textSecondary"
-                    gutterBottom
-                  >
-                    Trainer profile
-                  </Typography>
-                  <Typography variant="h5" component="h2" color="primary" align="left">
-                    {props.trainerData.name}
-                  </Typography>
-          
-                  <img
-                    className={classes.pic}
-                    src={props.trainerData.avatar}
-                    alt="Smiley face"
-                  />
-          
-                  {/* <Avatar aria-label="recipe" className={classes.avatar}></Avatar> */}
-          
-                  <Typography className={classes.pos} variant="h5" color="Secondary">
-                    Focus: {props.trainerData.about}
-                  </Typography>
-                  <Typography className={classes.desc} variant="body2" component="p">
-                    {props.trainerData.experience}
-                  </Typography>
-                </CardContent>
-                <CardActions>
-                  <Button size="small">update profile</Button>
-                </CardActions>
-              </Card>
-              </Paper>
-            </Grid>
-          </Grid>
-          <Box pt={4}>
-            <Copyright />
-          </Box>
-        </Container>
-      </main>
-      
-    </div>
-    )
-  }
+    <div className={classes.root} > 
+    <CssBaseline />
+    <TrainerAppSideBar />
+    <main className={classes.content}>
+      <div className={classes.appBarSpacer} />
+      <Container maxWidth="lg" className={classes.container}>
+        <Grid container spacing={10}>
+          {/* Students Images */}
+          <Grid item xs={12} md={12} lg={20}>
+            <Paper className={fixedHeightPaper}>
+              <ProfileTr trainerData={props.trainerData} />
+            </Paper>
+          </Grid>          
+        </Grid>
+        <Box pt={4}>
+          <Copyright />
+        </Box>
+      </Container>
+    </main>
+  </div>
 
+  );
+}
